python/oauth2-fastapi/pyproject.toml
[tool.poetry]
name = "oauth2-fastapi"
version = "0.1.0"
description = ""
authors = ["Felix Andreas <felix.andreas95@googlemail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
python-multipart = "^0.0.5"
tortoise = "^0.1.1"
tortoise-orm = "^0.17.2"
passlib = "^1.7.4"
PyJWT = "^2.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^20.8b1", allow-prereleases = true}
rope = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "dev:start"
